set up demo routes
